+++ src/server/index.ts
@@ -1,7 +1,7 @@
 import chalk from 'chalk';
 import { exit } from "process";
-import { Plugin, ViteDevServer } from "vite";
-import { RequestHandler, PLUGIN_NAME, ViteConfig, VitePluginNodeConfig, RequestHandlerOption } from "..";
+import { Connect, Plugin, ViteDevServer } from "vite";
+import { RequestAdapter, PLUGIN_NAME, ViteConfig, VitePluginNodeConfig, RequestAdapterOption } from "..";
 import { createDebugger } from '../utils';
 import { ExpressHandler } from "./express";
 import { NestHandler } from "./nest";
@@ -29,45 +29,40 @@ export const getPluginConfig = (server: ViteDevServer): VitePluginNodeConfig => 
   return (plugin.config!({}, { command: 'serve', mode: '' }) as ViteConfig).VitePluginNodeConfig
 }
 
-const getRequestHandler = (handler: RequestHandlerOption): RequestHandler | undefined => {
-  let requestHandler: RequestHandler | undefined
+const getRequestHandler = (handler: RequestAdapterOption): RequestAdapter | undefined => {
+  let requestHandler: RequestAdapter | undefined
 
   if (typeof handler === 'function') {
     debugServer(chalk.dim`server config set to custom`);
     requestHandler = handler;
   } else {
     debugServer(chalk.dim`creating ${handler} node server`);
-    requestHandler = SUPPORTED_FRAMEWORKS[handler] as RequestHandler;
+    requestHandler = SUPPORTED_FRAMEWORKS[handler] as RequestAdapter;
   }
 
   return requestHandler
 }
 
-export const startServer = async (server: ViteDevServer): Promise<void> => {
+export const createMiddleware = (server: ViteDevServer): Connect.HandleFunction => {
   const config = getPluginConfig(server);
   const logger = server.config.logger;
-  const requestHandler = getRequestHandler(config.handler);
+  const requestHandler = getRequestHandler(config.adapter);
 
   if (!requestHandler) {
     console.error('Failed to find a request handler');
     process.exit(1)
   }
 
-  const httpServer = http.createServer(async (req, res) => {
+  return async function (req: http.IncomingMessage, res: http.ServerResponse): Promise<void> {
     const appModule = await server.ssrLoadModule(config.appPath);
     let app = appModule[config.exportName!];
     if (!app) {
-      console.error(`Failed to find a named export ${config.exportName} from ${config.appPath}`);
+      logger.error(`Failed to find a named export ${config.exportName} from ${config.appPath}`);
       process.exit(1)
     } else {
       // some app may be created with a function returning a promise
       app = await app
       await requestHandler(app, req, res);
     }
-  });
-
-  httpServer.listen(config.server, () => {
-    logger.info(chalk.greenBright`Server started on http://${config.server?.host}:${config.server?.port}`);
-  });
+  }
 }
-